requirements.txt
Django>=5.0.0,<6.0.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.6
redis>=5.0.0
django-redis>=5.4.0
django-tailwind>=3.6.0
requests>=2.31.0
Pillow>=10.0.0
celery>=5.3.0
python-dotenv>=1.0.0
django-filter>=23.5
django-crispy-forms>=2.1
crispy-tailwind>=0.5.0
whitenoise>=6.5.0
dj-database-url>=2.1.0
djangorestframework>=3.15.0
